Re-organized files and iterated on some details
