+++ lib/golden_runner.dart
@@ -1,3 +1,4 @@
-export 'src/commands.dart';
+export 'src/commands/clean.dart';
+export 'src/commands/commands.dart';
 export 'src/docker.dart';
 export 'src/logging.dart';
